requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
cffi==1.14.6
cryptography==3.4.8
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework-jwt==1.11.0
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
requests==2.26.0
PyJWT==2.3.0